# docs/swagger.yaml
definitions:
...
  controllers.productInput:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      photo_path:
        type: string
      price:
        type: integer
      shop_id:
        type: integer
      stock:
        type: integer
      weight:
        type: integer
    type: object
  controllers.shopInput:
    properties:
      bank:
        type: string
      name:
        type: string
      phone:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
      photo_path:
        type: string
      price:
        type: integer
      shop_id:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
      weight:
        type: integer
    type: object
  models.Shop:
    properties:
      bank:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /categories/{id}/products:
    get:
      description: Get all Products by CategoryId
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get Products
      tags:
      - Public
...
  /products:
    get:
      description: Get a list of Product
      parameters:
      - description: name search by keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get all Product
      tags:
      - Public
    post:
      description: Creating new Product
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create new Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.productInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Create new Product
      tags:
      - Seller
      - Admin
  /products/{id}:
    delete:
      description: Delete product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete product
      tags:
      - Seller
      - Admin
    get:
      description: Get Product by Id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Get Product detail
      tags:
      - Public
    put:
      description: Update Product by id
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.productInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      summary: Update Product
      tags:
      - Seller
      - Admin
...
  /shops:
    get:
      description: Get a list of Shop
      parameters:
      - description: name search by keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Shop'
            type: array
      summary: Get all Shop
      tags:
      - Public
    post:
      description: Creating new Shop
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create new Shop
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.shopInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Shop'
      summary: Create new Shop
      tags:
      - Seller
      - Admin
  /shops/{id}:
    delete:
      description: Delete shop by id
      parameters:
      - description: Shop id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete shop
      tags:
      - Seller
      - Admin
    get:
      description: Get Shop by Id
      parameters:
      - description: Shop id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Shop'
      summary: Get Shop detail
      tags:
      - Public
    put:
      description: Update Shop by id
      parameters:
      - description: Shop id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update shop
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.shopInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Shop'
      summary: Update Shop
      tags:
      - Seller
      - Admin
  /shops/{id}/products:
    get:
      description: Get all Products by ShopId
      parameters:
      - description: Shop id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: Get Products
      tags:
      - Public
...
swagger: "2.0"